# 3.5.1 ## build: 依赖 ``` spring.boot.version > 2.5.4 mybatisplus.version > 3.4.3.1 antisamy.version>1.6.4 hutool.version>5.7.10 ```
## feat: 新增
- seata 1.4.2 的配置
- 用户管理页面，头像回显功能完善
- 新增可配置选项： 在执行sql时，忽略 租户插件自动拼接租户编码的表 (仅 COLUMN 模式有效)
- 新增可配置选项：分页插件支持更多的配置参数

## refactor： 增强
- 注释完善
- 默认的Controller分页查询，支持更多的扩展参数

## fix: 修复
- 实时获取文件访问地址失败时，忽略报错
- token 过期后，统一返回401状态码，使页面自动跳转到重新登录
- 排除slf4j冲突jar，解决控制台启动警告

+++ lamp-authority/lamp-authority-biz/src/main/java/top/tangyh/lamp/authority/service/auth/UserService.java
@@ -193,5 +193,5 @@ public interface UserService extends SuperCacheService<User>, LoadService {
      *
      * @return
      */
-    Integer todayUserCount();
+    Long todayUserCount();
 }
